requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
idna==2.7
isort==4.3.4
joblib==0.13.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
pylint==2.1.1
python-linkedin==4.1
python-slugify==1.2.6
requests==2.20.1
requests-oauthlib==1.0.0
selenium==3.141.0
six==1.11.0
SQLAlchemy==1.2.13
tinydb==3.12.0
typed-ast==1.1.0
Unidecode==1.0.22
urllib3==1.24.1
wrapt==1.10.11
